Prepare epoll implementation for rebinding of interfaces
To support on-the-fly interface deletion and recreation the way interfaces
are tracked has been changed.
Sockets are now defined through a specification and an instance. The
specification contains setup parameters for the socket instance.
The socket instance is bound to the current ifindex of its interface.
When the ifindex changes a special callback is invoked. The socket can
then be recreated/rebound to the new ifindex.